backend/requirements.txt
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-jwt
django-rest-auth[with_social]
django-celery-beat
django-celery-results
django-modeltranslation
django-extensions

gunicorn

psycopg2-binary

redis
celery

python-google-places
simplejson
